Login works but sessions don't seem to. Still can't get it to switch to the profile page after login.
