+++ controllers/homeRoutes.js
@@ -2,6 +2,7 @@ const router = require("express").Router();
 // const { Project, User } = require('../models');
 const withAuth = require("../utils/auth");
 
+
 const rp = require("request-promise");
 
 // + Array holding hardcoded version of the top tags for the main page to display when a user visits our website
@@ -149,36 +150,6 @@ router.get("/project/:id", async (req, res) => {
 	}
 });
 
-// Use withAuth middleware to prevent access to route
-router.get('/profile', withAuth, async (req, res) => {
-  try {
-    // Find the logged in user based on the session ID
-    const userData = await User.findByPk(req.session.user_id, {
-    //   attributes: { exclude: ['password'] },
-    //   include: [{ model: Project }],
-     });
 
-    const user = userData.get({ plain: true });
-
-    res.render('profile'
-    // , {
-    //   ...user,
-    //   logged_in: true
-    // }
-    );
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get("/login", (req, res) => {
-	// If the user is already logged in, redirect the request to another route
-	if (req.session.logged_in) {
-		res.redirect("/profile");
-		return;
-	}
-
-	res.render("login");
-});
 
 module.exports = router;
