+++ package.json
@@ -20,6 +20,8 @@
     "express-handlebars": "^5.2.0",
     "express-session": "^1.17.1",
     "mysql2": "^2.2.5",
+    "request": "^2.88.2",
+    "request-promise": "^4.2.6",
     "sequelize": "^6.3.5"
   },
   "devDependencies": {
