+++ views/profile.handlebars
@@ -9,4 +9,4 @@
   <button class="niche_btn">Niche 6</button>
 </div>
 
-<script src="/js/color.js"></script>
+<script src="../public/js/color.js"></script>
