+++ views/register.handlebars
@@ -22,4 +22,4 @@
     <button class="btn-large custom_btn login_btn">Sign Up!</button> 
 </form>
     
-{{!-- <script src="./js/login.js"></script> --}}
+<script src="../public/js/login.js"></script>
